Moved Xliff modules
