+++ extension/src/LanguageFunctions.ts
@@ -7,9 +7,9 @@ import * as VSCodeFunctions from './VSCodeFunctions';
 import * as escapeStringRegexp from 'escape-string-regexp';
 import { XliffIdToken } from './ALObject/XliffIdToken';
 import { Settings, Setting } from "./Settings";
-import { targetStateActionNeededToken, targetStateActionNeededKeywordList } from "./XlfFunctions";
+import { targetStateActionNeededToken, targetStateActionNeededKeywordList } from "./Xliff/XlfFunctions";
 import * as Logging from './Logging';
-import { CustomNoteType, Target, TargetState, TranslationToken, TransUnit, Xliff } from './XLIFFDocument';
+import { CustomNoteType, Target, TargetState, TranslationToken, TransUnit, Xliff } from './Xliff/XLIFFDocument';
 import { isNull, isNullOrUndefined } from 'util';
 import { BaseAppTranslationFiles, localBaseAppTranslationFiles } from './externalresources/BaseAppTranslationFiles';
 import { readFileSync } from 'fs';
