+++ extension/src/Xliff/XLIFFInterface.ts
